requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# Database
supabase>=2.0.0
redis>=5.0.0
pgvector>=0.3.0

# Code parsing
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-go>=0.23.0
tree-sitter-rust>=0.23.0

# Git operations
gitpython>=3.1.40

# Graph analysis
networkx>=3.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.0.0
structlog>=24.0.0
tiktoken>=0.7.0

# Async support
aiofiles>=23.0.0

# Type hints
typing-extensions>=4.0.0

